refactor contrib tests
